Move SiteHandler to site/
